sample-content-repo/fitness_tracker/backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
click==8.1.7